backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.12
email-validator==2.2.0
requests==2.32.3
httpx==0.27.2
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
alembic==1.13.3

# DICOM validation
pydicom==2.4.4

# Cache
redis==5.1.1

# Rate limiting
slowapi==0.1.9

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# Export
openpyxl==3.1.5

# Monitoring (optional)
prometheus-fastapi-instrumentator==7.0.0
